fixes for the https://t.me/c/1235155926/33801 commands

+++ pyrobot/plugins/admemes/whois.py
@@ -9,7 +9,7 @@ from pyrobot import COMMAND_HAND_LER
 from pyrobot.helper_functions.extract_user import extract_user
 
 
-@Client.on_message(Filters.command("whois", COMMAND_HAND_LER))
+@Client.on_message(Filters.command(["whois", "info", "id"], COMMAND_HAND_LER))
 async def who_is(client, message):
     """ extract user information """
     status_message = await message.reply_text(
@@ -36,16 +36,24 @@ async def who_is(client, message):
         message_out_str += f"Last Name: {from_user.last_name}\n"
         message_out_str += f"DC ID: <code>{from_user.dc_id}</code>\n"
         chat_photo = from_user.photo
-        local_user_photo = await client.download_media(
-            message=chat_photo.big_file_id
-        )
-        await message.reply_photo(
-            photo=local_user_photo,
-            quote=True,
-            caption=message_out_str,
-            parse_mode="html",
-            # ttl_seconds=,
-            disable_notification=True
-        )
-        os.remove(local_user_photo)
+        if chat_photo:
+            local_user_photo = await client.download_media(
+                message=chat_photo.big_file_id
+            )
+            await message.reply_photo(
+                photo=local_user_photo,
+                quote=True,
+                caption=message_out_str,
+                parse_mode="html",
+                # ttl_seconds=,
+                disable_notification=True
+            )
+            os.remove(local_user_photo)
+        else:
+            await message.reply_text(
+                text=message_out_str,
+                quote=True,
+                parse_mode="html",
+                disable_notification=True
+            )
         await status_message.delete()
